Shzhen/context awareness (#2337)
* refactor, bug to fix

* fix save userstate bug

* refactor: move turncontext in resolver

* merge

* AR for email skill (save list<recipient>)

* AR for email skill (save list<recipient>) 2

* AR for calendar

* refactor: split into 2 actions, rename
